feat: add stories, and move components
